pyproject.toml
[project]
name = "com"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sena-nana", email = "34237511+sena-nana@users.noreply.github.com" }
]
dependencies = [
    "pyserial>=3.5",
    "matplotlib>=3.8.2",
]
readme = "README.md"
requires-python = ">= 3.10"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
